requirements.txt
torch
torchvision
Pillow
requests